Draft: split Facebinder from Draft.py
