+++ src/Mod/Draft/CMakeLists.txt
@@ -69,6 +69,7 @@ SET(Draft_make_functions
     draftmake/make_circle.py
     draftmake/make_clone.py
     draftmake/make_ellipse.py
+    draftmake/make_facebinder.py
     draftmake/make_line.py
     draftmake/make_polygon.py
     draftmake/make_point.py
@@ -86,6 +87,7 @@ SET(Draft_objects
     draftobjects/circle.py
     draftobjects/clone.py
     draftobjects/ellipse.py
+    draftobjects/facebinder.py
     draftobjects/orthoarray.py
     draftobjects/polararray.py
     draftobjects/arc_3points.py
@@ -106,6 +108,7 @@ SET(Draft_view_providers
     draftviewproviders/view_base.py
     draftviewproviders/view_circulararray.py
     draftviewproviders/view_clone.py
+    draftviewproviders/view_facebinder.py
     draftviewproviders/view_orthoarray.py
     draftviewproviders/view_polararray.py
     draftviewproviders/view_draft_annotation.py
