+++ src/Mod/Draft/Draft.py
@@ -246,6 +246,13 @@ if FreeCAD.GuiUp:
     from draftviewproviders.view_point import ViewProviderPoint
     from draftviewproviders.view_point import _ViewProviderPoint
 
+# facebinder
+from draftmake.make_facebinder import make_facebinder, makeFacebinder
+from draftobjects.facebinder import Facebinder, _Facebinder
+if FreeCAD.GuiUp:
+    from draftviewproviders.view_facebinder import ViewProviderFacebinder
+    from draftviewproviders.view_facebinder import _ViewProviderFacebinder
+
 # working plane proxy
 from draftmake.make_wpproxy import make_workingplaneproxy
 from draftmake.make_wpproxy import makeWorkingPlaneProxy
@@ -2001,23 +2008,6 @@ def heal(objlist=None,delete=True,reparent=True):
         for n in dellist:
             FreeCAD.ActiveDocument.removeObject(n)
 
-def makeFacebinder(selectionset,name="Facebinder"):
-    """makeFacebinder(selectionset,[name]): creates a Facebinder object from a selection set.
-    Only faces will be added."""
-    if not FreeCAD.ActiveDocument:
-        FreeCAD.Console.PrintError("No active document. Aborting\n")
-        return
-    if not isinstance(selectionset,list):
-        selectionset = [selectionset]
-    fb = FreeCAD.ActiveDocument.addObject("Part::FeaturePython",name)
-    _Facebinder(fb)
-    if gui:
-        _ViewProviderFacebinder(fb.ViewObject)
-    faces = []
-    fb.Proxy.addSubobjects(fb,selectionset)
-    select(fb)
-    return fb
-
 
 def upgrade(objects,delete=False,force=None):
     """upgrade(objects,delete=False,force=None): Upgrades the given object(s) (can be
@@ -3732,104 +3722,6 @@ class _ShapeString(_DraftObject):
         glyphfaces.extend(islands)
         ret = Part.Compound(glyphfaces)           # should we fuse these instead of making compound?
         return ret
-
-
-class _Facebinder(_DraftObject):
-    """The Draft Facebinder object"""
-    def __init__(self,obj):
-        _DraftObject.__init__(self,obj,"Facebinder")
-        obj.addProperty("App::PropertyLinkSubList","Faces","Draft",QT_TRANSLATE_NOOP("App::Property","Linked faces"))
-        obj.addProperty("App::PropertyBool","RemoveSplitter","Draft",QT_TRANSLATE_NOOP("App::Property","Specifies if splitter lines must be removed"))
-        obj.addProperty("App::PropertyDistance","Extrusion","Draft",QT_TRANSLATE_NOOP("App::Property","An optional extrusion value to be applied to all faces"))
-        obj.addProperty("App::PropertyBool","Sew","Draft",QT_TRANSLATE_NOOP("App::Property","This specifies if the shapes sew"))
-
-
-    def execute(self,obj):
-        import Part
-        pl = obj.Placement
-        if not obj.Faces:
-            return
-        faces = []
-        for sel in obj.Faces:
-            for f in sel[1]:
-                if "Face" in f:
-                    try:
-                        fnum = int(f[4:])-1
-                        faces.append(sel[0].Shape.Faces[fnum])
-                    except(IndexError,Part.OCCError):
-                        print("Draft: wrong face index")
-                        return
-        if not faces:
-            return
-        try:
-            if len(faces) > 1:
-                sh = None
-                if hasattr(obj,"Extrusion"):
-                    if obj.Extrusion.Value:
-                        for f in faces:
-                            f = f.extrude(f.normalAt(0,0).multiply(obj.Extrusion.Value))
-                            if sh:
-                                sh = sh.fuse(f)
-                            else:
-                                sh = f
-                if not sh:
-                    sh = faces.pop()
-                    sh = sh.multiFuse(faces)
-                if hasattr(obj,"Sew"):
-                    if obj.Sew:
-                        sh = sh.copy()
-                        sh.sewShape()
-                if hasattr(obj,"RemoveSplitter"):
-                    if obj.RemoveSplitter:
-                        sh = sh.removeSplitter()
-                else:
-                    sh = sh.removeSplitter()
-            else:
-                sh = faces[0]
-                if hasattr(obj,"Extrusion"):
-                    if obj.Extrusion.Value:
-                        sh = sh.extrude(sh.normalAt(0,0).multiply(obj.Extrusion.Value))
-                sh.transformShape(sh.Matrix, True)
-        except Part.OCCError:
-            print("Draft: error building facebinder")
-            return
-        obj.Shape = sh
-        obj.Placement = pl
-
-    def addSubobjects(self,obj,facelinks):
-        """adds facelinks to this facebinder"""
-        objs = obj.Faces
-        for o in facelinks:
-            if isinstance(o,tuple) or isinstance(o,list):
-                if o[0].Name != obj.Name:
-                    objs.append(tuple(o))
-            else:
-                for el in o.SubElementNames:
-                    if "Face" in el:
-                        if o.Object.Name != obj.Name:
-                            objs.append((o.Object,el))
-        obj.Faces = objs
-        self.execute(obj)
-
-
-class _ViewProviderFacebinder(_ViewProviderDraft):
-    def __init__(self,vobj):
-        _ViewProviderDraft.__init__(self,vobj)
-
-    def getIcon(self):
-        return ":/icons/Draft_Facebinder_Provider.svg"
-
-    def setEdit(self,vobj,mode):
-        import DraftGui
-        taskd = DraftGui.FacebinderTaskPanel()
-        taskd.obj = vobj.Object
-        taskd.update()
-        FreeCADGui.Control.showDialog(taskd)
-        return True
-
-    def unsetEdit(self,vobj,mode):
-        FreeCADGui.Control.closeDialog()
-        return False
 
 
 ## @}
